Function imagine::png::unfilter_decompressed_data

pub fn unfilter_decompressed_data<F>(
    header: IHDR, 
    decompressed: &mut [u8], 
    op: F
) -> Result<(), PngError> where
    F: FnMut(u32, u32, &[u8]), 

Given the header, decompressed buffer, and a per-pixel op, unfilters the data and passes each pixel output to the op as the unfiltering occurs.

Each call to the op gets |x, y, data| as arguments, where x and y are the position of the pixel data (relative to the top left), and data is a slice of bytes representing the unfiltered pixel value at that location. Bit-packed pixel data will be unpacked and have the callback called once per pixel, with the data in the lowest bits of a single byte.

The data is unfiltered in place, and also each filter byte is reset to the “no filter” setting as well. Thus, it’s perfectly fine to call this more than once on the same decompressed data if you just want to iterate the data a second time for some reason.

Failure

You are allowed to pass a decompressed buffer larger than just the decompressed data itself. The function will use only the correct number of bytes from the start of the buffer.

If you for some reasons give a decompressed data buffer that is too small then you’ll get an error (possibly after some amount of the unfiltering is done).