Struct imagine::Image

pub struct Image { /* private fields */ }

This struct abstracts the implementation of the actual image away. It’s currently using image-rs under the hood but that is an implementation detail.

Every creation or cloning of the image struct should be available in debug logs because memory allocation is a potential bottle neck.

Implementations

impl Image

pub fn new(width: u32, height: u32) -> Self

Creates a new empty image with the given size.

Example

use imagine::Image;

let my_image = Image::new(100, 200);
let (w, h): (u32, u32) = my_image.size().into();
assert_eq!(w, 100);
assert_eq!(h, 200);

pub fn new_from_color(width: u32, height: u32, color: &Color) -> Self

pub fn size(&self) -> Size

pub fn new_stamp(color: &Color, hardness: f64, radius: f64) -> Self

pub fn new_four_pixels(a: &str, b: &str, c: &str, d: &str) -> Self

Creates a 2x2 image using the given color strings (for test purposes)

+---+---+
| a | b |
+---+---+
| c | d |
+---+---+

pub fn from_file(path: &str) -> Result<Self, ImageError>

Creates a new image with the content of the given image file.

pub fn from_bytes(buffer: &[u8]) -> Result<Self, ImageError>

pub fn into_array(&self) -> &Vec<u8>

pub fn into_png_bytes(&self) -> Vec<u8>

pub fn from_base64(base64: &str) -> Result<Self, Box<dyn Error>>

pub fn save(&self, path: &str) -> Result<(), ImageError>

pub fn encode_base64(&self) -> Result<String, ImageError>

pub fn height(&self) -> u32

pub fn width(&self) -> u32

pub fn into_bytes(&self) -> Vec<u8>

pub fn pixel(&self, x: u32, y: u32) -> Color

pub fn put_pixel(&mut self, x: u32, y: u32, g: Color)

pub fn clean(&mut self, area: &Rectangle)

impl Image

pub fn grayscale(&mut self)

pub fn gaussian_noise(&mut self, mean: f64, stddev: f64, seed: u64)

pub fn flip_horizontally(&mut self)

pub fn flip_vertically(&mut self)

pub fn draw_line(&mut self, stamp: &Image, track: &[Position]) -> Rectangle

Draws the stamp image at each position of the track on the image struct.

Returns damaged area in image coordinates.